feat: added search vector client

+++ services/vector/src/main/java/com/baskaaleksander/nuvine/application/dto/TextVectorSearchRequest.java
@@ -20,6 +20,6 @@ public record TextVectorSearchRequest(
         @Size(min = 8, max = 14, message = "Top K must be between 8 and 14")
         int topK,
         @Size(min = 0, max = 1, message = "Threshold must be between 0 and 1")
-        long threshold
+        float threshold
 ) {
 }
